pyproject.toml
[project]
name = "endress-promag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pydoover>=0.4.13",
    "transitions>=0.9.2",
    "websockets",
    "pycryptodome",
]

[project.scripts]
doover-app-run = "endress_promag:main"
export-config = "endress_promag.app_config:export"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "aiohttp>=3.11.18",
    "grpcio>=1.71.0",
    "protobuf>=6.30.2",
    "pytest>=8.3.5",
    "pytest-asyncio>=0.26.0",
    "requests>=2.32.3",
]
